First pass at setting up automatic tests
