+++ setup.cfg
@@ -9,6 +9,12 @@ inplace = True
 zip_safe = True
 install_requires = 
   xxhash
+  numpy >=1.20
   cython
   sympy >=1.8
   typeguard
+  petsc4py
+  mpi4py
+  meshio
+  gmsh
+  pygmsh
