+++ tests/test_001_meshes.py
@@ -1,12 +1,11 @@
-import pytest
 
 def test_generate_2d_quads_mesh():
-    from underworld3 import mesh
+    from underworld3 import meshes
     resX, resY = 10, 10
     minX, maxX = -5.0, 5.0
     minY, maxY = -5.0, 5.0
 
-    mesh = mesh.Box(elementRes=(resX, resY), 
+    mesh = meshes.Box(elementRes=(resX, resY), 
                 minCoords=(minX, minY),
                 maxCoords=(maxX, maxY),
                 simplex=False)
@@ -15,13 +14,13 @@ def test_generate_2d_quads_mesh():
     assert mesh.data.shape == ((resX+1)*(resY+1), mesh.dim)
 
 def test_generate_3d_quads_mesh():
-    from underworld3 import mesh
+    from underworld3 import meshes
     resX, resY, resZ = 10, 10, 10
     minX, maxX = -5.0, 5.0
     minY, maxY = -5.0, 5.0
     minZ, maxZ = -5.0, 5.0
 
-    mesh = mesh.Box(elementRes=(resX, resY, resZ), 
+    mesh = meshes.Box(elementRes=(resX, resY, resZ), 
                 minCoords=(minX, minY, minZ),
                 maxCoords=(maxX, maxY, maxZ),
                 simplex=False)
@@ -30,12 +29,12 @@ def test_generate_3d_quads_mesh():
     assert mesh.data.shape == ((resX+1)*(resY+1)*(resZ+1), mesh.dim)
 
 def test_generate_2d_simplices_mesh():
-    from underworld3 import mesh
+    from underworld3 import meshes
     resX, resY = 10, 10
     minX, maxX = -5.0, 5.0
     minY, maxY = -5.0, 5.0
 
-    mesh = mesh.Box(elementRes=(resX, resY), 
+    mesh = meshes.Box(elementRes=(resX, resY), 
                 minCoords=(minX, minY),
                 maxCoords=(maxX, maxY),
                 simplex=True)
@@ -44,13 +43,13 @@ def test_generate_2d_simplices_mesh():
     assert mesh.data.shape == ((resX+1)*(resY+1), mesh.dim)
 
 def test_generate_3d_simplices_mesh():
-    from underworld3 import mesh
+    from underworld3 import meshes
     resX, resY, resZ = 10, 10, 10
     minX, maxX = -5.0, 5.0
     minY, maxY = -5.0, 5.0
     minZ, maxZ = -5.0, 5.0
 
-    mesh = mesh.Box(elementRes=(resX, resY, resZ), 
+    mesh = meshes.Box(elementRes=(resX, resY, resZ), 
                 minCoords=(minX, minY, minZ),
                 maxCoords=(maxX, maxY, maxZ),
                 simplex=True)
@@ -64,7 +63,7 @@ def test_generate_3d_simplices_mesh():
 #    mesh = Spherical(refinements=4)
 #    return
 
-def test_mesh_save(load_multi_meshes):
-    load_multi_meshes.save("mesh.h5")
-    return
+#def test_mesh_save(load_multi_meshes):
+#    load_multi_meshes.save("mesh.h5")
+#    return
