+++ tests/test_101_kdtree.py
@@ -1,8 +1,20 @@
 import underworld3 as uw
 import numpy as np
-from petsc4py import PETSc
+import pytest
+
+
+test_single_data = []
+
+coords = np.zeros((1,2))
+coords[0] = (0.5,)*2
+test_single_data.append((100000,2,coords))
+
+coords = np.zeros((1,3))
+coords[0] = (0.5,)*3
+test_single_data.append((100000,3,coords))
 
 # Single coord test
+@pytest.mark.parametrize("n,dim,coords", test_single_data)
 def test_single_coord(n,dim,coords):
     """
     Create set of randomly distributed points. 
@@ -23,23 +35,13 @@ def test_single_coord(n,dim,coords):
     dot2 = np.sum(np.multiply(diff,diff),1)
     brutept = np.argmin(dot2)
 
-    if found[0] != True:
-        raise RuntimeError("KDTree did not appear to find a point close to coord.")
-    if kdpt[0] != brutept:
-        raise RuntimeError("KDTree and brute force method did not find the same point.")
-    if not np.allclose(dist[0],dot2[brutept]):
-        raise RuntimeError( "KDTree and Numpy did not find the same distance squared.\n"\
-                           f"KDTree distance={dist[0]} Numpy distance={dot2[brutept]} ")
-
-coords = np.zeros((1,2))
-coords[0] = (0.5,)*2
-test_single_coord(100000,2,coords)
-coords = np.zeros((1,3))
-coords[0] = (0.5,)*3
-test_single_coord(100000,3,coords)
+    assert found[0], "KDTree did not appear to find a point close to coord."
+    assert kdpt[0] == brutept, "KDTree and brute force method did not find the same point."
+    assert np.allclose(dist[0],dot2[brutept]), "KDTree and Numpy did not find the same distance squared.\n"\
+                           f"KDTree distance={dist[0]} Numpy distance={dot2[brutept]} "
 
 
-# Self test
+@pytest.mark.parametrize("n,dim", [(10000, 2), (10000, 3)])
 def test_self_points(n,dim):
     """
     Create set of randomly distributed points. 
@@ -56,19 +58,15 @@ def test_self_points(n,dim):
     # Use KDTree to find closest point to a coord
     (kdpt, dist, found) = index.find_closest_point(pts)
 
-    if not np.allclose(True, found):
-        raise RuntimeError("All points should have been found.")
+    assert np.allclose(True, found), "All points should have been found."
     # `find_closest_point` should return index of pts.
-    if not np.allclose(np.arange(n),kdpt):
-        raise RuntimeError("Point indices weren't as expected.")
+    assert np.allclose(np.arange(n),kdpt), "Point indices weren't as expected."
     # Distance to self should be zero!
-    if not np.allclose(0.,dist):
-        raise RuntimeError("Point distances weren't as expected.")
+    assert np.allclose(0.,dist), "Point distances weren't as expected."
 
-test_self_points(10000,2)
-test_self_points(10000,3)
 
 # Mesh vertex test
+@pytest.mark.parametrize("res,dim", [(16, 2), (8, 3)])
 def test_mesh_verts(res, dim):
     """
     Generate a mesh and create the kd index
@@ -77,7 +75,7 @@ def test_mesh_verts(res, dim):
     and confirm displaced coords are still closest
     to mesh verts.
     """
-    mesh = uw.mesh.Box(elementRes=(res,)*dim)
+    mesh = uw.meshes.Box(elementRes=(res,)*dim)
     index = uw.algorithms.KDTree(mesh.data)
     index.build_index()
 
@@ -87,22 +85,17 @@ def test_mesh_verts(res, dim):
     elsize = 1./float(res)
     coords = mesh.data.copy() + 0.5*elsize*np.random.random( mesh.data.shape )
     (kdpt, dist, found) = index.find_closest_point(coords)
-    if not np.allclose(True, found):
-        raise RuntimeError("All points should have been found.")
+    assert np.allclose(True, found), "All points should have been found."
     # `find_closest_point` should return index of pts.
-    if not np.allclose(np.arange(mesh.data.shape[0]),kdpt):
-        raise RuntimeError("Point indices weren't as expected.")
+    assert np.allclose(np.arange(mesh.data.shape[0]),kdpt), "Point indices weren't as expected."
     # Calc distances
     diff = mesh.data - coords
     dot2 = np.sum(np.multiply(diff,diff),1)
-    if not np.allclose(dot2,dist):
-        raise RuntimeError("Point distances weren't as expected.")
-
-test_mesh_verts(16,2)
-test_mesh_verts(8,3)
+    assert np.allclose(dot2,dist), "Point distances weren't as expected."
 
 
 # Mesh centroid test
+@pytest.mark.parametrize("res,dim,ppcell", [(16, 2, 4), (16, 3, 4)])
 def test_mesh_centroid(res,dim,ppcell):
     """
     Generate a mesh and create the kd index
@@ -111,7 +104,7 @@ def test_mesh_centroid(res,dim,ppcell):
     that closest centroid coincides with particle 
     owning cell. 
     """
-    mesh = uw.mesh.Box(elementRes=(res,)*dim)
+    mesh = uw.meshes.Box(elementRes=(res,)*dim)
     centroids = np.zeros((res**dim,dim))
     for index in range(res**dim):
         centroids[index] = mesh.dm.computeCellGeometryFVM(index)[1]
@@ -125,11 +118,7 @@ def test_mesh_centroid(res,dim,ppcell):
     with swarm.access():
         kdpt, dist, found = index.find_closest_point(swarm.data)
 
-        if not np.allclose(True, found):
-            raise RuntimeError("All points should have been found.")
+        assert np.allclose(True, found), "All points should have been found."
         # `find_closest_point` should return index of pts.
-        if not np.allclose(swarm.particle_cellid.data[:,0],kdpt):
-            raise RuntimeError("Point indices weren't as expected.")
+        assert np.allclose(swarm.particle_cellid.data[:,0],kdpt), "Point indices weren't as expected."
 
-test_mesh_centroid(16,2,4)
-test_mesh_centroid(16,3,4)
